requirements.txt
Django>=2.2.1,<2.3.0
django-private-storage>=2.2,<2.3
djangorestframework>=3.9.2,<3.10.0
djangorestframework-jwt
django-rest-auth